requirements.txt
datasets
prefixspan
google-generativeai
python-docx
